完善自动建表语法，删除无用代码

+++ src/main/java/com/ctgu/entity/User.java
@@ -36,8 +36,7 @@ public class User implements Serializable
 
 	private Date updateTime;
 
-	private Byte isDelete;
-	
+	private Integer isDelete;
 
 	public Integer getId()
 	{
@@ -169,12 +168,12 @@ public class User implements Serializable
 		this.updateTime = updateTime;
 	}
 
-	public Byte getIsDelete()
+	public Integer getIsDelete()
 	{
 		return isDelete;
 	}
 
-	public void setIsDelete(Byte isDelete)
+	public void setIsDelete(Integer isDelete)
 	{
 		this.isDelete = isDelete;
 	}
